look ay my machine learn

+++ lib/MNIST/MNIST_Image.h
@@ -10,7 +10,7 @@ using namespace std;
 class MNIST_Image
 {
 public:
-    uint8_t label;
+    int label;
     uint16_t pixels[784];
     uint16_t size = 784;
     MNIST_Image(string line);
